+++ core/utils.py
@@ -1,0 +1,15 @@
+
+
+
+def format_time(duration_ms):
+    """格式化歌曲时长"""
+    duration = duration_ms // 1000
+
+    hours = duration // 3600
+    minutes = (duration % 3600) // 60
+    seconds = duration % 60
+
+    if hours > 0:
+        return f"{hours}:{minutes:02d}:{seconds:02d}"
+    else:
+        return f"{minutes:02d}:{seconds:02d}"
